// frontend/generated/models/BrandDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 8lind8box Platform API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface BrandDto
 */
export interface BrandDto {
    /**
     * 
     * @type {number}
     * @memberof BrandDto
     */
    brandId?: number;
    /**
     * 
     * @type {string}
     * @memberof BrandDto
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof BrandDto
     */
    description?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BrandDto
     */
    isVisible?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof BrandDto
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof BrandDto
     */
    updatedAt?: Date;
}

/**
 * Check if a given object implements the BrandDto interface.
 */
export function instanceOfBrandDto(value: object): value is BrandDto {
    return true;
}

export function BrandDtoFromJSON(json: any): BrandDto {
    return BrandDtoFromJSONTyped(json, false);
}

export function BrandDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BrandDto {
    if (json == null) {
        return json;
    }
    return {
        
        'brandId': json['brandId'] == null ? undefined : json['brandId'],
        'name': json['name'] == null ? undefined : json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'isVisible': json['isVisible'] == null ? undefined : json['isVisible'],
        'createdAt': json['createdAt'] == null ? undefined : (new Date(json['createdAt'])),
        'updatedAt': json['updatedAt'] == null ? undefined : (new Date(json['updatedAt'])),
    };
}

export function BrandDtoToJSON(json: any): BrandDto {
    return BrandDtoToJSONTyped(json, false);
}

export function BrandDtoToJSONTyped(value?: BrandDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'brandId': value['brandId'],
        'name': value['name'],
        'description': value['description'],
        'isVisible': value['isVisible'],
        'createdAt': value['createdAt'] == null ? undefined : ((value['createdAt']).toISOString()),
        'updatedAt': value['updatedAt'] == null ? undefined : ((value['updatedAt']).toISOString()),
    };
}